Rebase

+++ package/batocera/core/batocera-configgen/configgen/configgen/generators/dolphin/dolphinControllers.py
@@ -3,6 +3,7 @@
 
 import batoceraFiles
 import os
+import codecs
 from Emulator import Emulator
 from utils.logger import eslog
 
@@ -52,13 +53,17 @@ def generateControllerConfig_emulatedwiimotes(playersControllers, rom):
     extraOptions["Source"] = "1"
 
     # side wiimote
+    # l2 for shaking actions
     if ".side." in rom:
         extraOptions["Options/Sideways Wiimote"] = "1"
         wiiMapping['x']   = 'Buttons/B'
         wiiMapping['y'] = 'Buttons/A'
         wiiMapping['a']   = 'Buttons/2'
         wiiMapping['b'] = 'Buttons/1'
-        
+        wiiMapping['l2'] = 'Shake/X'
+        wiiMapping['l2'] = 'Shake/Y'
+        wiiMapping['l2'] = 'Shake/Z'
+
 
     # i: infrared, s: swing, t: tilt, n: nunchuk
     # 12 possible combinations : is si / it ti / in ni / st ts / sn ns / tn nt
@@ -122,7 +127,7 @@ def generateControllerConfig_emulatedwiimotes(playersControllers, rom):
         wiiMapping['joystick2up'] = 'Classic/Right Stick/Up'
         wiiMapping['joystick2left'] = 'Classic/Right Stick/Left'
 
-    #This section allows a per ROM override of the default key options.  
+    #This section allows a per ROM override of the default key options.
     configname = rom + ".cfg"  #Define ROM configuration name
     if os.path.isfile(configname): #file exists
         import ast
@@ -165,7 +170,7 @@ def generateControllerConfig_gamecube(playersControllers,rom):
         'joystick1down': 'down',
         'joystick1right': 'right'
     }
-    
+
     #This section allows a per ROM override of the default key options.
     configname = rom + ".cfg"  #Define ROM configuration name
     if os.path.isfile(configname): #file exists
@@ -183,11 +188,12 @@ def generateControllerConfig_gamecube(playersControllers,rom):
 
 def removeControllerConfig_gamecube():
     configFileName = "{}/{}".format(batoceraFiles.dolphinConfig, "GCPadNew.ini")
-    os.remove(configFileName)
+    if os.path.isfile(configFileName):
+        os.remove(configFileName)
 
 def generateControllerConfig_realwiimotes(filename, anyDefKey):
     configFileName = "{}/{}".format(batoceraFiles.dolphinConfig, filename)
-    f = open(configFileName, "w")
+    f = codecs.open(configFileName, "w", encoding="utf_8_sig")
     nplayer = 1
     while nplayer <= 4:
         f.write("[" + anyDefKey + str(nplayer) + "]" + "\n")
@@ -198,7 +204,7 @@ def generateControllerConfig_realwiimotes(filename, anyDefKey):
 
 def generateHotkeys(playersControllers):
     configFileName = "{}/{}".format(batoceraFiles.dolphinConfig, "Hotkeys.ini")
-    f = open(configFileName, "w")
+    f = codecs.open(configFileName, "w", encoding="utf_8_sig")
 
     hotkeysMapping = {
         'a':  'Keys/Reset',                      'b': 'Keys/Toggle Pause',
@@ -214,7 +220,7 @@ def generateHotkeys(playersControllers):
     for playercontroller, pad in sorted(playersControllers.items()):
         if nplayer == 1:
             f.write("[Hotkeys1]" + "\n")
-            f.write("Device = evdev/0/" + pad.realName + "\n")
+            f.write("Device = evdev/0/" + pad.realName.strip() + "\n")
 
             # search the hotkey button
             hotkey = None
@@ -245,7 +251,7 @@ def generateHotkeys(playersControllers):
 
 def generateControllerConfig_any(playersControllers, filename, anyDefKey, anyMapping, anyReverseAxes, anyReplacements, extraOptions = {}):
     configFileName = "{}/{}".format(batoceraFiles.dolphinConfig, filename)
-    f = open(configFileName, "w")
+    f = codecs.open(configFileName, "w", encoding="utf_8_sig")
     nplayer = 1
     nsamepad = 0
 
@@ -261,7 +267,7 @@ def generateControllerConfig_any(playersControllers, filename, anyDefKey, anyMap
         double_pads[pad.configName] = nsamepad+1
 
         f.write("[" + anyDefKey + str(nplayer) + "]" + "\n")
-        f.write("Device = evdev/" + str(nsamepad) + "/" + pad.realName + "\n")
+        f.write("Device = evdev/" + str(nsamepad).strip() + "/" + pad.realName.strip() + "\n")
         for opt in extraOptions:
             f.write(opt + " = " + extraOptions[opt] + "\n")
